gazetteer_drescher v0.1.0 GazetteerDrescher.Writing

This module handles all writing to the output file.

Functions

open_output_file(file)

Creates and opens the file at the specified location.

Returns: {:ok, io_device}, see http://elixir-lang.org/docs/stable/elixir/File.html#open!/2

write_place(place)

Chooses transformation module based on requested output format and hands over the given place data (JSON parsed as Map). Then writes the transformed data to file.

Returns: :ok